requirements.txt
requests
python-dotenv
pydantic
firecrawl-py
prediction-market-agent-tooling[langchain]
black
mypy
